requirements.txt
streamlit==1.44.1
streamlit-cropper==0.2.2
streamlit-drawable-canvas==0.9.3
transformers==4.51.3
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
scikit-learn
matplotlib
seaborn
wandb
Pillow==11.0.0
sympy==1.13.1
tenacity==9.1.2
toml==0.10.2
evaluate==0.4.3
environ==1.0
accelerate>=0.26.0
gdown
